Moved all binding and old input tests into BindingTest and OldInput test classes.
